Refactor history.go into service and constant files
