+++ consent-server/internal/consent/constants.go
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com).
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package consent
+
+// HistoryReason is the server-generated reason attached to consent amendment history.
+type HistoryReason string
+
+const (
+	HistoryReasonConsentAmended                        HistoryReason = "Consent amended"
+	HistoryReasonConsentDetailsAmended                 HistoryReason = "Consent details amended"
+	HistoryReasonConsentAttributesAmended              HistoryReason = "Consent attributes amended"
+	HistoryReasonConsentAuthorizationsAmended          HistoryReason = "Consent authorizations amended"
+	HistoryReasonConsentPurposesAmended                HistoryReason = "Consent purposes amended"
+	HistoryReasonConsentRevoked                        HistoryReason = "Consent revoked"
+	HistoryReasonConsentExpired                        HistoryReason = "Consent expired"
+	HistoryReasonConsentDetailsAmendedAndReactivated   HistoryReason = "Consent details amended and reactivated"
+	HistoryReasonConsentAuthorizationsAmendedAndStatus HistoryReason = "Consent authorizations amended and status updated"
+)
